refactor: extract exhaustiveness module from checker
Move match exhaustiveness checking from checker/match_check.rs into a
standalone exhaustiveness.rs module that operates as a pure function.

The new module takes subject type, match arms, span, and a type
resolver callback, returning Vec<Diagnostic>. This separates
exhaustiveness analysis from the checker's mutable state, allowing it
to evolve independently (e.g., decision tree compilation for codegen).

Pattern type checking (check_pattern) stays in the checker since it
needs environment access for variable bindings.

Closes #1110
